% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyTSA-package.R
\docType{package}
\name{easyTSA-package}
\alias{easyTSA}
\alias{easyTSA-package}
\title{easyTSA: Trial Sequential Analysis for meta objects}
\description{
Runs Trial Sequential Analysis (TSA, Copenhagen Trial Unit) on
meta-analysis objects from the \pkg{meta} package, draws the TSA graph with
\pkg{ggplot2}, and exchanges \code{.TSA} files with the original Java program
(installed separately from \url{https://ctu.dk/tsa/}).
}
\section{Guides (read in order)}{

\enumerate{
\item \link{easyTSA-1-workflow} -- from a \code{metabin} object to a TSA plot.
\item \link{easyTSA-2-theory} -- what the RIS, boundaries and inner wedge are.
\item \link{easyTSA-3-plot} -- customising the plot.
\item \link{easyTSA-4-software} -- opening the analysis in the TSA program.
}
}

\section{Main functions}{

\describe{
\item{\code{\link[=tsa_create]{tsa_create()}}}{Run the TSA on a \code{metabin}/\code{metacont} object.}
\item{\code{\link[=tsa_plot]{tsa_plot()}}}{Draw the TSA graph (\code{ggplot}).}
\item{\code{\link[=summary]{summary()}}}{Per-look table and TSA-adjusted CI.}
\item{\code{\link[=tsa_write]{tsa_write()}}, \code{\link[=tsa_read]{tsa_read()}}}{Exchange \code{.TSA} files.}
\item{\code{\link[=tsa_launch]{tsa_launch()}}}{Start the TSA software (needs Java and the program).}
}
}

\section{Building blocks}{

\describe{
\item{\code{\link[=tsa_ris]{tsa_ris()}}, \code{\link[=tsa_diversity]{tsa_diversity()}}}{Required information size, D-squared.}
\item{\code{\link[=tsa_bounds]{tsa_bounds()}}, \code{\link[=tsa_futility]{tsa_futility()}}, \code{\link[=tsa_spending]{tsa_spending()}}}{Lan-DeMets
boundaries and spending functions.}
}
}

\section{Example data}{

\describe{
\item{\link{atb_peecs}}{Prophylactic antibiotics vs none for post-ESD
electrocoagulation syndrome (7 studies).}
}
}

\examples{
m <- meta::metabin(event.e, n.e, event.c, n.c, data = atb_peecs,
                   studlab = paste(author, year), sm = "RR",
                   method = "MH", method.tau = "REML",
                   random = TRUE, common = FALSE)
x <- tsa_create(m, rrr = 0.30, label.e = "ATB", label.c = "No ATB",
                title = "PEECS after ESD")
summary(x)
tsa_plot(x)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vanioljantunes/easyTSA}
  \item Report bugs at \url{https://github.com/vanioljantunes/easyTSA/issues}
}

}
\author{
\strong{Maintainer}: Vanio Antunes \email{vanio.antunes.md@gmail.com}

Authors:
\itemize{
  \item Vanio Antunes \email{vanio.antunes.md@gmail.com}
}

}
\keyword{internal}
