% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{tsa_read}
\alias{tsa_read}
\title{Read a .TSA file into a data frame}
\usage{
tsa_read(file)
}
\arguments{
\item{file}{Path to a \code{.TSA} file.}
}
\value{
A data frame with \code{studlab}, \code{year} and either
\verb{event.e, n.e, event.c, n.c} or \verb{n.e, mean.e, sd.e, n.c, mean.c, sd.c}.
Attribute \code{"settings"} holds the meta-analysis block as a named list.
}
\description{
Parses the trials of a \code{.TSA} file (TSA program format) into a data
frame ready for \code{\link[meta:metabin]{meta::metabin()}}. Settings of the meta-analysis block
are returned as attributes.
}
\examples{
f <- system.file("extdata", "atb_peecs.TSA", package = "easyTSA")
d <- tsa_read(f)
d
attr(d, "settings")$identifier
}
